+++ fmpy/fmi1.py
@@ -169,6 +169,8 @@ class _FMU(object):
                 if isinstance(v, c_void_p):
                     v = v.value
                 a += hex(0 if v is None else v)
+            elif t == c_bool:
+                a += str(v.value)
             elif t == POINTER(c_uint):
                 # value references
                 if v is None:
@@ -500,7 +502,7 @@ class FMU1Slave(_FMU1):
         return self.fmi1CancelStep(self.component)
 
     def doStep(self, currentCommunicationPoint, communicationStepSize, newStep=fmi1True):
-        return self.fmi1DoStep(self.component, currentCommunicationPoint, communicationStepSize, newStep)
+        self.fmi1DoStep(self.component, currentCommunicationPoint, communicationStepSize, newStep)
 
     def getStatus(self, kind):
         value = fmi1Status(fmi1OK)